SDK for python
Python constants are created using the QuickTemplate build pipeline.

The SSE generator is a class with `classmethod`s which perform the
various SDK functions. The use of classmethods will allow users to
extend the class if needed. (Thanks to Tom D.
<noreplygitemail@protonmail.com> for the suggestion for this, plus the
use of `__slots__`.)

Also provides helpers and examples for the following frameworks:

* Sanic
* FastAPI
* Django
* Quart

pyproject updates as suggested by dalito