Support nested regex and strings in expressions
This regex allows Datastar expressions to support nested
regex and strings that contain ; and/or \n without breaking.

Full regex for testing on regex101.com

```
((?:\/(?:\\\/|[^\/])*\/|"(?:\\"|[^\"])*"|'(?:\\'|[^'])*'|`(?:\\`|[^`])*`|[^;\n])*)([;\n]+)
```

Only valid ; and \n should be in the second capture group Each of these regex ignore a block type:

```
regex            \/(?:\\\/|[^\/])*\/
double quotes     "(?:\\"|[^\"])*"
single quotes     '(?:\\'|[^'])*'
ticks             `(?:\\`|[^`])*`
```
We want to ignore the non delimiter part of statements too:

`[^;\n]`

Once all the blocks we want to ignore are captured we then find
the statement delimiters in the second capture group:

`([;\n]+)`

Note: the reason why this regex isn't broken down into strings is that
doing so adds an extra layer of escapes which makes it even harder to
reason about, harder to check in external regex apps and leads to
worse compression.

The ignore block model allows us to extend this and validate each
component separately.

Effectively, an ignore block is a regex that skips the content of the
block. For example we want to skip the contents of a `''` block then
we need to match the open quote and the close quote and ignore the
contents including escaped quotes.

That's the theory anyway. You can go mad testing this, just remember
the it needs to be valid JS! A few times I thought I'd broken
it and it would have broken the JS parser.

I don't expect this to be perfect, but I do expect it to be easy to
fix any edge cases we find by design (what I'm aiming for with this).

There might be a combination of nested strings that break this.

I've run the test (sort is still failing but it was before) and
I've built and tested the todo app.