Typescript sdk standardization (#700)
* Bring typescript sdk in line with sdk api's

But still allow passing of stricter types when it makes sense

* fix bug introduced in https://github.com/starfederation/datastar/pull/698

we want the stream to be ended only if keepalive is not present

* Remove close method from typescript sdk and improve comments

The stream is closed automatically after the callback runs.
The only situation where we want to use keepalive is if we
want an infinite stream, in which case we dont want to close it.

* improve package.json for typescript sdk

* Use consts VERSION in typescript sdk tests

* Improve method comments in typescript sdk

this sdk is also used in javascript where there's less type information