Expand the attribute casing documentation
The previous top-line message "data-* attributes are case-insensitive", while
true, was very confusing for new users because it is not immedietly clear that
sentence is referring to the HTML spec and not the library of the docs they are reading.

So I've clarified the top-line message, and then expanded the explanation in a
way I hope makes it easier to understand the first time.