Improve casing and modifier docs (#779)
* Emphasize two of the most important facts about d* attributes.

* Expand the attribute casing documentation

The previous top-line message "data-* attributes are case-insensitive", while
true, was very confusing for new users because it is not immedietly clear that
sentence is referring to the HTML spec and not the library of the docs they are reading.

So I've clarified the top-line message, and then expanded the explanation in a
way I hope makes it easier to understand the first time.

* Explain how to use modifiers with `data-signals`

(the object notation variant)

* Clarify it is possible to set both data-signals__ifmissing= and data-signals on the same element

* Update attribute_plugins.md

* Update attribute_plugins.md

---------

Co-authored-by: Ben Croker <57572400+bencroker@users.noreply.github.com>