remove panic call from one ExecuteScriptOption
Changed the shape of the ExecuteScriptOption option to return an error,
which allows validation of the Option parameters. This allowed removing
a panic call from one of the options.

Panic calls are common in the standard library. However, Datastar calls
return errors. Mixing panics and error returns in the same construct is
slight code smell. More importantly, the mixing produces a less
intuitive and predictable API.